Cloud Function requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.6.0
msal==1.21.0
pandas==1.5.3
pyarrow==11.0.0
openpyxl==3.1.2
requests==2.28.2
